+++ tests/performance/logistic.cpp
@@ -1,10 +1,11 @@
 #include <charmtyles/charmtyles.hpp>
+#include <string>
 
 #include "logistic.decl.h"
 
 #include <Eigen/Dense>
 
-/* readonly */ int DIMENSION;
+/* readonly */ size_t DIMENSION;
 
 // features -> X_train (Data Points * Num Features)
 // target -> y_train (Data Points)
@@ -110,13 +111,13 @@ public:
     {
         DIMENSION = 1 << 14;
         if (msg->argc > 1)
-            DIMENSION = atoi(msg->argv[1]);
+            DIMENSION = std::stoull(msg->argv[1]);
 
         ct::init();
         thisProxy.benchmark(DIMENSION);
     }
 
-    void benchmark(int dim)
+    void benchmark(size_t dim)
     {
         std::shared_ptr<training_generator> new_training_generator =
             std::make_shared<training_generator>();
